backend/requirements.txt
fastapi[standard]
httpx
websockets
uvicorn
pydantic-settings
